adds alternate source for sqs-specific aws credentials
When using this plugin with other hubot plugins it might be useful or
necessary to use a separate set of AWS credentials that have the
specific SQS permissions needed. Currently this plugin uses the default
credential configuration supported by aws-sdk (environment variable or
credentials stored in a file like ~/.aws/credentials).

This PR adds an additional method of providing alternate credentials
solely for SQS, by defining these variables:

- HUBOT_SQS_AWS_ACCESS_KEY_ID
- HUBOT_SQS_SECRET_ACCESS_KEY

If these are defined, they'll be passed into the AWS.SQS instance and
used for communication with SQS. If not defined, the default aws-sdk
credential locations will be checked.

This naming convention is similar to how other packages like
hubot-s3-brain name their credential environment variables.

This commit also modifies the README and CHANGELOG.md in anticipation of
a future release.